requirements.txt
Flask==1.1.2
google-auth==1.22.0
Pillow==7.2.0
requests==2.24.0
urllib3==1.25.10
